more general commit history (#198)
