+++ cargo-smart-release/src/lib.rs
@@ -1,6 +1,7 @@
 pub use context::Context;
 
 pub mod command;
+pub mod commit;
 mod context;
 pub mod git;
 pub mod traverse;
